refactor(tools): 将 MCP hooks 和组件迁入 tools 目录完成页面内聚
