+++ src/renderer/pages/settings/tools/components/AddMcpServerModal.tsx
