+++ src/renderer/pages/settings/tools/components/McpAgentStatusDisplay.tsx
