+++ src/renderer/pages/settings/tools/hooks/useMcpOAuth.ts
